+++ tests/fkooman/Rest/ServiceTest.php
@@ -761,12 +761,29 @@ class ServiceTest extends PHPUnit_Framework_TestCase
         );
         $request = new Request("http://www.example.org/index.php", "GET");
         $response = $service->run($request);
-        $this->assertEquals(301, $response->getStatusCode());
+        $this->assertEquals(302, $response->getStatusCode());
         $this->assertEquals("http://www.example.org/index.php/manage/", $response->getHeader('Location'));
         $request = new Request("http://www.example.org/index.php", "GET");
         $request->setPathInfo('/manage/');
         $response = $service->run($request);
         $this->assertEquals(200, $response->getStatusCode());
         $this->assertEquals('default_route_works', $response->getContent());
+    }
+
+    /**
+     * @expectedException fkooman\Http\Exception\NotFoundException
+     * @expectedExceptionMessage url not found
+     */
+    public function testNoPathInfo()
+    {
+        $service = new Service();
+        $service->get(
+            '/foo',
+            function () {
+                return "foo";
+            }
+        );
+        $request = new Request("http://www.example.org/index.php", "GET");
+        $service->run($request);
     }
 }